Former-commit-id: 72dc71812429fe5e8e8f831f1da98d86712dcdb7

+++ plugin/bullet/lib/bullet2/src/vectormath/sse/vec_aos.h
@@ -577,7 +577,12 @@ VECTORMATH_FORCE_INLINE const Vector3 Vector3::operator -( ) const
 {
 	//return Vector3(_mm_sub_ps( _mm_setzero_ps(), mVec128 ) );
 
-	VM_ATTRIBUTE_ALIGN16 static const int array[] = {0x80000000, 0x80000000, 0x80000000, 0x80000000};
+	VM_ATTRIBUTE_ALIGN16 static const int array[] = {
+        static_cast<int>(0x80000000),
+        static_cast<int>(0x80000000),
+        static_cast<int>(0x80000000),
+        static_cast<int>(0x80000000)
+    };
 	__m128 NEG_MASK = SSEFloat(*(const vec_float4*)array).vf;
 	return Vector3(_mm_xor_ps(get128(),NEG_MASK));
 }
